Largest Number
Write a function that takes three number arguments as input and finds the largest of them. Print the following text on the console:  `The largest number is {number}.`.
The input comes as three number arguments passed to your function.
The output should be printed to the console.

Example:
Input: 5, -3, 16  Output: The largest number is 16.
Input: -3, -5, -22.5  Output: The largest number is -3.